+++ CHANGELOG.md
@@ -13,6 +13,10 @@ The format is based on [Keep a Changelog][keepachangelog] and this project adher
 - Minimal required `monolog/monolog` version now is `^2.0`
 - `DefaultLogstashLogger`, `DefaultUdpLogstashLogger`, `EventsLogstashLogger` and `EventsUdpLogstashLogger` uses `'extra'` for extra keys (instead `false`) and `'context'` for context keys (instead `null`) inside logstash "fields" by default
 
+### Added
+
+- Method `skipLogging(): bool` in `ShouldBeLoggedContract` interface
+
 ## v2.2.0
 
 ### Changed